Bonsoir

+++ ft_strlen.c
@@ -6,14 +6,16 @@
 /*   By: ibouabda <ibouabda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/06 08:11:52 by ibouabda          #+#    #+#             */
-/*   Updated: 2018/11/06 08:32:30 by ibouabda         ###   ########.fr       */
+/*   Updated: 2018/11/26 15:42:22 by ibouabda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include "libft.h"
+
+size_t	ft_strlen(char *str)
 {
 	char	c;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	c = *str;
